Erste Versuche mit XCS gemacht. Funktioniert noch nicht komplett. Nächste Schritte: Einführung von GA. Maximale Anzahl von Classifiern einfügen. Beobachtung muss überarbeitet werden. Lokalerer Ansatz.

+++ PacMan/src/game/entries/pacman/group4/IRewarder.java
@@ -24,7 +24,7 @@ public interface IRewarder {
 	 * @param reward
 	 *            entsprechender Reward für ein erreichtes Ziel.
 	 */
-	public void giveRewardToActions(int reward);
+	public void giveRewardToActions(double reward);
 
 	/**
 	 * Alle Classifier, die nicht im ActionSet sind müssen Taxes bezahlen.
